Added x11 unicode input support. Added wayland support (FTE_TARGET=linux-wayland). Added properly support for rpi+egl (FTE_TARGET=linux-rpi, warning: expects to cross-compile).
git-svn-id: https://svn.code.sf.net/p/fteqw/code/trunk@4622 fc73d0e0-1445-4013-8a0c-d673dee63da5

+++ engine/common/fs.c
@@ -590,11 +590,11 @@ void COM_Locate_f (void)
 	{
 		if (!*loc.rawname)
 		{
-			Con_Printf("File is %i bytes compressed inside %s\n", loc.len, loc.search->logicalpath);
+			Con_Printf("File is %u bytes compressed inside %s\n", (unsigned)loc.len, loc.search->logicalpath);
 		}
 		else
 		{
-			Con_Printf("Inside %s (%i bytes)\n  %s\n", loc.rawname, loc.len, loc.search->logicalpath);
+			Con_Printf("Inside %s (%u bytes)\n  %s\n", loc.rawname, (unsigned)loc.len, loc.search->logicalpath);
 		}
 	}
 	else
